Structure changes - Creating TeachmeProvider for managing all the calls to SDK from one place
