+++ src/app/components/layout/screens/information-screen/index.tsx
@@ -1,6 +1,6 @@
 import React, { useRef, useEffect } from "react";
 
-import localization from "../../../../consts/localization";
+import localization from "../../../../constants/localization";
 
 import { IInformationScreenData, InformationScreenType } from "./interface";
