+++ src/app/constants/localization.ts
