+++ src/index.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { StrictMode } from "react";
 import ReactDOM from "react-dom";
 
 import App from "./app/App";
@@ -6,4 +6,9 @@ import App from "./app/App";
 // global styles
 import "./styles/reset.less";
 
-ReactDOM.render(React.createElement(App), document.getElementById("root"));
+ReactDOM.render(
+  <StrictMode>
+    <App />
+  </StrictMode>,
+  document.getElementById("root")
+);
